+++ clox/src/debug.h
@@ -4,6 +4,6 @@
 #include "chunk.h"
 #include "utils/common.h"
 
-int disassembleInstruction(const Chunk& chunk, uint16_t offset, bool linesAvailable);
+uint16_t disassembleInstruction(const Chunk& chunk, uint16_t offset, bool linesAvailable);
 
 void disassemble(const Chunk& chunk, const char* name);
